+++ tests/gpu_megatron/torch/utils/plugins/test_utils_megatron.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
+import pytest
 from _test_utils.torch.megatron.models import get_mcore_qwen3_600m
 from _test_utils.torch.megatron.utils import initialize_for_megatron
 from transformers import AutoTokenizer
@@ -22,12 +22,18 @@ from modelopt.torch.utils.plugins import megatron_generate, megatron_mmlu
 
 SEED = 1234
 
+# TODO: move to regression test folder
 
-def _test_megatron_generate_and_mmlu(rank, size):
-    initialize_for_megatron(tensor_model_parallel_size=size, seed=SEED)
 
-    model = get_mcore_qwen3_600m(tensor_model_parallel_size=size).cuda().eval()
-
+def _test_megatron_generate_and_mmlu(rank, size, parallelism):
+    if parallelism == "tp":
+        initialize_for_megatron(tensor_model_parallel_size=size, seed=SEED)
+        model = get_mcore_qwen3_600m(tensor_model_parallel_size=size).cuda().eval()
+    elif parallelism == "pp":
+        initialize_for_megatron(pipeline_model_parallel_size=size, seed=SEED)
+        model = get_mcore_qwen3_600m(pipeline_model_parallel_size=size).cuda().eval()
+    else:
+        raise ValueError(f"Invalid parallelism: {parallelism}")
     tokenizer = AutoTokenizer.from_pretrained("Qwen/Qwen3-0.6B")
 
     messages = [
@@ -42,10 +48,13 @@ def _test_megatron_generate_and_mmlu(rank, size):
     model_inputs = tokenizer([text], return_tensors="pt").to(device="cuda")
     output_ids = megatron_generate(model, model_inputs["input_ids"])
     output_text = tokenizer.batch_decode(output_ids)
-    print(output_text)
+    print(rank, output_text)
 
-    assert megatron_mmlu(model, tokenizer) > 0.24
+    assert 0.36 < megatron_mmlu(model, tokenizer, fraction=0.1, batch_size=16) < 0.39
 
 
-def test_megatron_generate_and_mmlu(dist_workers):
-    dist_workers.run(_test_megatron_generate_and_mmlu)
+@pytest.mark.parametrize("parallelism", ["tp", "pp"])
+def test_megatron_generate_and_mmlu(dist_workers, parallelism, num_gpus):
+    if num_gpus == 1 and parallelism == "pp":
+        pytest.skip("Skipping as redundant test on 1 GPU")
+    dist_workers.run(_test_megatron_generate_and_mmlu, parallelism=parallelism)